refactor: Extracted convert table to json function to a dedicated file
